% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{coef.cv.hmcox}
\alias{coef.cv.hmcox}
\title{Coefficients of a cross-validated fit}
\usage{
\method{coef}{cv.hmcox}(object, rule = object$rule, ...)
}
\arguments{
\item{object}{a \code{cv.hmcox} object.}

\item{rule}{"min" or "1se"; defaults to the rule used when fitting.}

\item{...}{unused.}
}
\description{
Coefficients of a cross-validated fit
}
