% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcox.R
\name{hmc_lambda_seq}
\alias{hmc_lambda_seq}
\title{Lambda sequence for the global effects}
\usage{
hmc_lambda_seq(D, nlambda = 25, lambda.min.ratio = NULL)
}
\description{
Lambda sequence for the global effects
}
\keyword{internal}
